requirements.txt
python-telegram-bot
python-binance
tinydb
requests